split tracing init into unit
